+++ CHANGELOG.md
@@ -6,8 +6,9 @@ See updating [Changelog example here](https://keepachangelog.com/en/1.0.0/)
 ## [Unreleased]
 
 ### Added
-- loadbalancer: `WaitForLoadBalancerOperationalState` helper for waiting for the load balancer to achieve a desired operational state
+- loadbalancer: `WaitForLoadBalancerOperationalState` helper for waiting for the load balancer to achieve a desired operational state.
 - loadbalancer: `WaitForLoadBalancerDeletion` blocks execution until the specified load balancer instance has been deleted.
+- `GetDevicesAvailability` function for listing device availability per zone.
 
 ## [8.20.0]