+++ CHANGELOG.md
@@ -5,6 +5,8 @@ See updating [Changelog example here](https://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased]
 
+## [8.21.0]
+
 ### Added
 - loadbalancer: `WaitForLoadBalancerOperationalState` helper for waiting for the load balancer to achieve a desired operational state.
 - loadbalancer: `WaitForLoadBalancerDeletion` blocks execution until the specified load balancer instance has been deleted.
@@ -603,7 +605,8 @@ See updating [Changelog example here](https://keepachangelog.com/en/1.0.0/)
 
 First stable release
 
-[Unreleased]: https://github.com/UpCloudLtd/upcloud-go-api/compare/v8.20.0...HEAD
+[Unreleased]: https://github.com/UpCloudLtd/upcloud-go-api/compare/v8.21.0...HEAD
+[8.21.0]: https://github.com/UpCloudLtd/upcloud-go-api/compare/v8.20.0...v8.21.0
 [8.20.0]: https://github.com/UpCloudLtd/upcloud-go-api/compare/v8.19.0...v8.20.0
 [8.19.0]: https://github.com/UpCloudLtd/upcloud-go-api/compare/v8.18.0...v8.19.0
 [8.18.0]: https://github.com/UpCloudLtd/upcloud-go-api/compare/v8.17.0...v8.18.0